Split issue message conversions below file limit
